Add component ResultListCard
